Fixes release build errors due to misplaced #if DEBUG / #endif
